Enable cppcoreguidelines-narrowing-conversions
Make implicit narrowing conversions explicit, and fix the sites where
the narrowing was a real bug rather than silencing it with a cast:

- checker: validate opset_import versions fit in int
  (checked_opset_version) instead of silently truncating; a wrapped
  version could collide with the -1 "not found" sentinel and bypass
  opset compatibility checks.
- Drop wrong-direction casts where the destination is already int64_t
  (InferShapesImpl, graphProtoToGraph, GetNodeDeterminism).
- Col2Im: accumulate block_shape_size in int64_t; the int accumulator
  truncated each dim and could overflow.
- Window ops and MelWeightMatrix: convert output_datatype with the
  checked narrow<> helper so wrapped attribute values fail inference.
- Gather data propagation: narrow<> the propagated index value so
  out-of-range indices fail instead of wrapping into valid range.
- Einsum: reject output labels missing from the inputs (operator[]
  default-inserted index 0); retype the label map value to int.
- Pooling (opsets 1-11): compute ceil_mode output sizes with exact
  integer arithmetic instead of 32-bit float, matching current opsets;
  float rounding gave off-by-one dims beyond 2^24.
- narrow<>: take the value by value so lvalue arguments compile.
- MathOpTwoIntegers takes and returns int64_t rather than int; the int
  signature truncated dim values above INT_MAX.
- Inliner: convert domain_version once instead of casting at each use.

Enables the check via both its bugprone- and cppcoreguidelines- alias
names.

+++ onnx/defs/data_propagators.h
@@ -75,7 +75,7 @@ inline void GatherOp13DataPropagator(DataPropagationContext& ctx) {
   TensorShapeProto tsp;
   for (int i = 0; i < input_indices->dim_size(); ++i) {
     if (input_indices->dim(i).has_dim_value()) {
-      appendDimToTensorShapeProto(tsp, input_data, input_indices->dim(i).dim_value());
+      appendDimToTensorShapeProto(tsp, input_data, narrow<int>(input_indices->dim(i).dim_value()));
     } else {
       return;
     }
